backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
rdkit-pypi
propka
scipy
numpy
python-multipart
